packages/billy/pyproject.toml
[project]
name = "billy"
version = "0.2.0"
description = "BILLY (Internal Diagnostic Tool)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "anthropic>=0.21.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "asyncpg>=0.29.0",
    "PyJWT[crypto]>=2.8.0",
]

[project.scripts]
billy = "billy.server:main"
billy-dev = "billy.dev:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/billy"]
